Add categories, improve users

+++ users/users.py
@@ -1,11 +1,10 @@
-from datetime import timedelta
-from typing import List
+from typing import List, Optional
 
-from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 
-from . import crud, schemas, auth
-from dependencies import get_db,  oauth2_request_form
+from . import crud, schemas
+from dependencies import get_db, paginator, permissions
 from .auth import get_current_user
 
 router = APIRouter(
@@ -15,8 +14,9 @@ router = APIRouter(
 
 
 @router.get('/', response_model=List[schemas.User])
-def read_users(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
-    users = crud.get_users(db, skip=skip, limit=limit)
+def read_users(page: int = 0, db: Session = Depends(get_db)):
+    pagination = paginator.paginate(page)
+    users = crud.get_users(db, pagination=pagination)
     return users
 
 
@@ -41,8 +41,24 @@ def read_user(user_id: int, db: Session = Depends(get_db)):
     return db_user
 
 
+@router.patch('/{user_id}', response_model=schemas.User)
+def change_user(user_id: int, user: schemas.UserPatch,
+                current_user: schemas.User = Depends(get_current_user),
+                db: Session = Depends(get_db)):
+    if current_user.permissions >= permissions.get('admin') or current_user.id == user_id:
+        db_user = crud.update_user(db, user_id, user)
+        if db_user is None:
+            raise HTTPException(status_code=404, detail="Uporabnik ne obstaja")
+        return db_user
+    else:
+        raise HTTPException(status_code=403, detail="Nimate dovoljenja")
+
+
 @router.delete('/{user_id}')
-def delete_user(user_id: int, db: Session = Depends(get_db)):
-    if not crud.delete_user(db, user_id):
-        raise HTTPException(status_code=404, detail="Uporabnik ne obstaja")
-    return {"message": "Izbris uspešen"}
+def delete_user(user_id: int, current_user: schemas.User = Depends(get_current_user), db: Session = Depends(get_db)):
+    if current_user.permissions >= permissions.get('admin') or current_user.id == user_id:
+        if not crud.delete_user(db, user_id):
+            raise HTTPException(status_code=404, detail="Uporabnik ne obstaja")
+        return {"detail": "Izbris uspešen"}
+    else:
+        raise HTTPException(status_code=403, detail="Nimate dovoljenja")
